[MZ] Some refactor to remove variable name clashes

+++ mz_frontend/SDLVersion/sdl2.c
@@ -34,7 +34,7 @@ void pickOrDrop();
 
 void pickItem();
 
-void graphicsPut(int16_t x, int16_t y, uint8_t colour) {
+void graphicsPut(int16_t x, int16_t y, uint16_t colour) {
     if (x < 0) {
         x = 0;
     }
@@ -59,7 +59,7 @@ void graphicsPut(int16_t x, int16_t y, uint8_t colour) {
 #endif
 }
 
-void hLine(int16_t x0, int16_t x1, int16_t y, uint8_t colour) {
+void hLine(int16_t x0, int16_t x1, int16_t y, uint16_t colour) {
     if (y < 0) {
         return;
     }
@@ -85,7 +85,7 @@ void hLine(int16_t x0, int16_t x1, int16_t y, uint8_t colour) {
     }
 }
 
-void vLine(int16_t x0, int16_t y0, int16_t y1, uint8_t colour) {
+void vLine(int16_t x0, int16_t y0, int16_t y1, uint16_t colour) {
 
     if (x0 < 0) {
         return;
@@ -127,7 +127,7 @@ void clearGraphics() {
     memset(framebuffer, 0, 256 * 160);
 }
 
-void writeStr(uint8_t nColumn, uint8_t nLine, char *str, uint8_t fg, uint8_t bg) {
+void writeStr(uint8_t nColumn, uint8_t nLine, char *str, uint16_t fg, uint16_t bg) {
     puts(str);
 }
 
@@ -164,7 +164,7 @@ void pickItem();
 void clearScreen() {}
 
 
-uint8_t getKey() {
+uint16_t getKey() {
     SDL_Event event;
 
     mBufferedCommand = '.';
@@ -381,7 +381,7 @@ void HUD_initialPaint() {
 void HUD_refresh() {
 
 
-    for (uint8_t i = 0; i < 6; ++i) {
+    for (uint16_t i = 0; i < 6; ++i) {
         writeStr(16, 13 + i, (i == cursorPosition) ? ">" : " ", 2, 0);
     }
